PythonProject-master/requirements.txt
fastapi
uvicorn
sqlalchemy
aiomysql
python-dotenv
alembic
pyjwt
python-jose
python-multipart
passlib
bcrypt
cryptography